Simplify property finding.
